adds copyright headers, extracts ArrayConverter to top level class
